Backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.12.1
django-environ==0.12.0
djangorestframework==3.16.1
environ==1.0
gunicorn==23.0.0
idna==3.11
jwt==1.4.0
packaging==25.0
pycparser==2.23
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0